+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes between versions
 
+## Not yet released
+
+* Fixed compatibility with Windows Subsystem for Linux
+
 ## 2.0.1 (2018-04-04)
 
 * Fixed autoloader in CLI script